% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extraCols_broadPeak}
\alias{extraCols_broadPeak}
\title{broadPeak extraCols}
\format{A named character vector with 3 values:
\describe{
  \item{signalValue}{"numeric"}
  \item{pValue}{"numeric"}
  \item{qValue}{"numeric"}
}}
\source{
\url{https://charlesjb.github.io/How_to_import_narrowPeak/}
}
\usage{
extraCols_broadPeak
}
\description{
The extraCols value to import broadPeak files with rtracklayer.
}
\keyword{datasets}
