% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extraCols_narrowPeak}
\alias{extraCols_narrowPeak}
\title{narrowPeak extraCols}
\format{A named character vector with 4 values:
\describe{
  \item{signalValue}{"numeric"}
  \item{pValue}{"numeric"}
  \item{qValue}{"numeric"}
  \item{peak}{"integer"}
}}
\source{
\url{https://charlesjb.github.io/How_to_import_narrowPeak/}
}
\usage{
extraCols_narrowPeak
}
\description{
The extraCols values to import narrowPeak files with rtracklayer.
}
\keyword{datasets}
