% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_file}
\alias{get_demo_file}
\title{Get the file path to a specific file format demo file.}
\usage{
get_demo_file(format = c("bed", "narrowPeak", "broadPeak"))
}
\arguments{
\item{format}{The file format to return. Default: "bed"}
}
\value{
The path to the file corresponding to the specified file format.
}
\description{
Available format:
    * bed
    * narrowPeak
    * broadPeak
}
\examples{
get_demo_file() # Return path to bed file by default
get_demo_file(format = "narrowPeak")

}
