% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_peaks}
\alias{import_peaks}
\title{Import peaks}
\usage{
import_peaks(filename)
}
\arguments{
\item{filename}{A vector with the path of the file to import.}
}
\value{
A list of GRanges.
}
\description{
Import narrowPeak or broadPeak files
}
\examples{
filename <- get_demo_file("narrowPeak")
gr <- import_peaks(filename)

}
